Refactor FinaidMailer and RecuploadMailer templates for improved clarity and user engagement
This commit removes default sender emails from the FinaidMailer and RecuploadMailer classes, enhancing the clarity of the email templates. It also introduces new text email templates for the RecuploadMailer, ensuring consistent communication with users. Additionally, the HTML templates are updated to include personalized greetings and direct links to the student portal, improving user experience and engagement.